@@ -26,6 +26,7 @@ const (
 	ErrorUnknownIndexType
 	ErrorEmptyID
 	ErrorIndexReadInconsistency
+	ErrorTwoPhaseSearchInconsistency
 )
 
 // Error represents a more strongly typed bleve error for detecting
@@ -37,14 +38,15 @@ func (e Error) Error() string {
 }
 
 var errorMessages = map[Error]string{
-	ErrorIndexPathExists:        "cannot create new index, path already exists",
-	ErrorIndexPathDoesNotExist:  "cannot open index, path does not exist",
-	ErrorIndexMetaMissing:       "cannot open index, metadata missing",
-	ErrorIndexMetaCorrupt:       "cannot open index, metadata corrupt",
-	ErrorIndexClosed:            "index is closed",
-	ErrorAliasMulti:             "cannot perform single index operation on multiple index alias",
-	ErrorAliasEmpty:             "cannot perform operation on empty alias",
-	ErrorUnknownIndexType:       "unknown index type",
-	ErrorEmptyID:                "document ID cannot be empty",
-	ErrorIndexReadInconsistency: "index read inconsistency detected",
+	ErrorIndexPathExists:             "cannot create new index, path already exists",
+	ErrorIndexPathDoesNotExist:       "cannot open index, path does not exist",
+	ErrorIndexMetaMissing:            "cannot open index, metadata missing",
+	ErrorIndexMetaCorrupt:            "cannot open index, metadata corrupt",
+	ErrorIndexClosed:                 "index is closed",
+	ErrorAliasMulti:                  "cannot perform single index operation on multiple index alias",
+	ErrorAliasEmpty:                  "cannot perform operation on empty alias",
+	ErrorUnknownIndexType:            "unknown index type",
+	ErrorEmptyID:                     "document ID cannot be empty",
+	ErrorIndexReadInconsistency:      "index read inconsistency detected",
+	ErrorTwoPhaseSearchInconsistency: "2-phase search failed, likely due to an overlapping topology change",
 }
