@@ -16,10 +16,11 @@ package scorch
 
 import (
 	"fmt"
+	"sync/atomic"
+
 	"github.com/RoaringBitmap/roaring"
 	index "github.com/blevesearch/bleve_index_api"
 	segment "github.com/blevesearch/scorch_segment_api/v2"
-	"sync/atomic"
 )
 
 var OptimizeConjunction = true
