@@ -21,6 +21,8 @@ import (
 
 	"github.com/blevesearch/bleve/v2/mapping"
 	"github.com/blevesearch/bleve/v2/search"
+	"github.com/blevesearch/bleve/v2/search/collector"
+	"github.com/blevesearch/bleve/v2/search/query"
 	index "github.com/blevesearch/bleve_index_api"
 )
 
@@ -160,13 +162,88 @@ func (i *indexAliasImpl) SearchInContext(ctx context.Context, req *SearchRequest
 	if len(i.indexes) < 1 {
 		return nil, ErrorAliasEmpty
 	}
+	if _, ok := ctx.Value(search.PreSearchKey).(bool); ok {
+		// since presearchKey is set, it means that the request
+		// is being executed as part of a presearch, which
+		// indicates that this index alias is set as an Index
+		// in another alias, so we need to do a presearch search
+		// and NOT a real search
+		return preSearchDataSearch(ctx, req, i.indexes...)
+	}
+
+	// at this point we know we are doing a real search
+	// either after a presearch is done, or directly
+	// on the alias
+
+	// check if request has preSearchData which would indicate that the
+	// request has already been preSearched and we can skip the
+	// preSearch step now, we call an optional function to
+	// redistribute the preSearchData to the individual indexes
+	// if necessary
+	var preSearchData map[string]map[string]interface{}
+	if req.PreSearchData != nil {
+		if requestHasKNN(req) {
+			var err error
+			preSearchData, err = redistributeKNNPreSearchData(req, i.indexes)
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
 
 	// short circuit the simple case
 	if len(i.indexes) == 1 {
+		if preSearchData != nil {
+			req.PreSearchData = preSearchData[i.indexes[0].Name()]
+		}
 		return i.indexes[0].SearchInContext(ctx, req)
 	}
 
-	return MultiSearch(ctx, req, i.indexes...)
+	// at this stage we know we have multiple indexes
+	// check if preSearchData needs to be gathered from all indexes
+	// before executing the query
+	var err error
+	// only perform presearch if
+	//  - the request does not already have preSearchData
+	//  - the request requires presearch
+	var preSearchDuration time.Duration
+	var sr *SearchResult
+	if req.PreSearchData == nil && preSearchRequired(req) {
+		searchStart := time.Now()
+		preSearchResult, err := preSearch(ctx, req, i.indexes...)
+		if err != nil {
+			return nil, err
+		}
+		// check if the presearch result has any errors and if so
+		// return the search result as is without executing the query
+		// so that the errors are not lost
+		if preSearchResult.Status.Failed > 0 {
+			return preSearchResult, nil
+		}
+
+		// if there are no errors, then merge the data in the presearch result
+		preSearchResult = mergePreSearchResult(req, preSearchResult, i.indexes)
+		if requestSatisfiedByPreSearch(req) {
+			sr = finalizeSearchResult(req, preSearchResult)
+			// no need to run the 2nd phase MultiSearch(..)
+		} else {
+			preSearchData, err = constructPreSearchData(req, preSearchResult, i.indexes)
+			if err != nil {
+				return nil, err
+			}
+		}
+		preSearchDuration = time.Since(searchStart)
+	}
+
+	// check if search result was generated as part of presearch itself
+	if sr == nil {
+		sr, err = MultiSearch(ctx, req, preSearchData, i.indexes...)
+		if err != nil {
+			return nil, err
+		}
+	}
+	sr.Took += preSearchDuration
+	return sr, nil
 }
 
 func (i *indexAliasImpl) Fields() ([]string, error) {
@@ -429,8 +506,8 @@ func (i *indexAliasImpl) Swap(in, out []Index) {
 // the actual final results.
 // Perhaps that part needs to be optional,
 // could be slower in remote usages.
-func createChildSearchRequest(req *SearchRequest) *SearchRequest {
-	return copySearchRequest(req)
+func createChildSearchRequest(req *SearchRequest, preSearchData map[string]interface{}) *SearchRequest {
+	return copySearchRequest(req, preSearchData)
 }
 
 type asyncSearchResult struct {
@@ -439,9 +516,209 @@ type asyncSearchResult struct {
 	Err    error
 }
 
+func preSearchRequired(req *SearchRequest) bool {
+	return requestHasKNN(req)
+}
+
+func preSearch(ctx context.Context, req *SearchRequest, indexes ...Index) (*SearchResult, error) {
+	// create a dummy request with a match none query
+	// since we only care about the preSearchData in PreSearch
+	dummyRequest := &SearchRequest{
+		Query: query.NewMatchNoneQuery(),
+	}
+	newCtx := context.WithValue(ctx, search.PreSearchKey, true)
+	if requestHasKNN(req) {
+		addKnnToDummyRequest(dummyRequest, req)
+	}
+	return preSearchDataSearch(newCtx, dummyRequest, indexes...)
+}
+
+func tagHitsWithIndexName(sr *SearchResult, indexName string) {
+	for _, hit := range sr.Hits {
+		hit.IndexNames = append(hit.IndexNames, indexName)
+	}
+}
+
+// if the request is satisfied by just the presearch result,
+// finalize the result and return it directly without
+// performing multi search
+func finalizeSearchResult(req *SearchRequest, preSearchResult *SearchResult) *SearchResult {
+	if preSearchResult == nil {
+		return nil
+	}
+
+	// global values across all hits irrespective of pagination settings
+	preSearchResult.Total = uint64(preSearchResult.Hits.Len())
+	maxScore := float64(0)
+	for i, hit := range preSearchResult.Hits {
+		// since we are now using the presearch result as the final result
+		// we can discard the indexNames from the hits as they are no longer
+		// relevant.
+		hit.IndexNames = nil
+		if hit.Score > maxScore {
+			maxScore = hit.Score
+		}
+		hit.HitNumber = uint64(i)
+	}
+	preSearchResult.MaxScore = maxScore
+	// now apply pagination settings
+	var reverseQueryExecution bool
+	if req.SearchBefore != nil {
+		reverseQueryExecution = true
+		req.Sort.Reverse()
+		req.SearchAfter = req.SearchBefore
+	}
+	if req.SearchAfter != nil {
+		preSearchResult.Hits = collector.FilterHitsBySearchAfter(preSearchResult.Hits, req.Sort, req.SearchAfter)
+	}
+	preSearchResult.Hits = hitsInCurrentPage(req, preSearchResult.Hits)
+	if reverseQueryExecution {
+		// reverse the sort back to the original
+		req.Sort.Reverse()
+		// resort using the original order
+		mhs := newSearchHitSorter(req.Sort, preSearchResult.Hits)
+		req.SortFunc()(mhs)
+		req.SearchAfter = nil
+	}
+
+	if req.Explain {
+		preSearchResult.Request = req
+	}
+	return preSearchResult
+}
+
+func mergePreSearchResult(req *SearchRequest, res *SearchResult,
+	indexes []Index) *SearchResult {
+	if requestHasKNN(req) {
+		res.Hits = mergeKNNDocumentMatches(req, res.Hits)
+	}
+	return res
+}
+
+func requestSatisfiedByPreSearch(req *SearchRequest) bool {
+	if requestHasKNN(req) && isKNNrequestSatisfiedByPreSearch(req) {
+		return true
+	}
+	return false
+}
+
+func constructPreSearchData(req *SearchRequest, preSearchResult *SearchResult, indexes []Index) (map[string]map[string]interface{}, error) {
+	mergedOut := make(map[string]map[string]interface{}, len(indexes))
+	for _, index := range indexes {
+		mergedOut[index.Name()] = make(map[string]interface{})
+	}
+	var err error
+	if requestHasKNN(req) {
+		mergedOut, err = constructKnnPresearchData(mergedOut, preSearchResult, indexes)
+		if err != nil {
+			return nil, err
+		}
+	}
+	return mergedOut, nil
+}
+
+func preSearchDataSearch(ctx context.Context, req *SearchRequest, indexes ...Index) (*SearchResult, error) {
+	asyncResults := make(chan *asyncSearchResult, len(indexes))
+
+	// run search on each index in separate go routine
+	var waitGroup sync.WaitGroup
+
+	var searchChildIndex = func(in Index, childReq *SearchRequest) {
+		rv := asyncSearchResult{Name: in.Name()}
+		rv.Result, rv.Err = in.SearchInContext(ctx, childReq)
+		asyncResults <- &rv
+		waitGroup.Done()
+	}
+
+	waitGroup.Add(len(indexes))
+	for _, in := range indexes {
+		go searchChildIndex(in, createChildSearchRequest(req, nil))
+	}
+
+	// on another go routine, close after finished
+	go func() {
+		waitGroup.Wait()
+		close(asyncResults)
+	}()
+
+	var sr *SearchResult
+	indexErrors := make(map[string]error)
+
+	for asr := range asyncResults {
+		if asr.Err == nil {
+			if sr == nil {
+				// first result
+				sr = asr.Result
+				tagHitsWithIndexName(sr, asr.Name)
+			} else {
+				// merge with previous
+				tagHitsWithIndexName(asr.Result, asr.Name)
+				sr.Merge(asr.Result)
+			}
+		} else {
+			indexErrors[asr.Name] = asr.Err
+		}
+	}
+
+	// merge just concatenated all the hits
+	// now lets clean it up
+
+	// handle case where no results were successful
+	if sr == nil {
+		sr = &SearchResult{
+			Status: &SearchStatus{
+				Errors: make(map[string]error),
+			},
+		}
+	}
+
+	// in presearch partial results are not allowed as it can lead to
+	// the real search giving incorrect results, and hence the search
+	// result is reset.
+	// discard partial hits if some child index has failed or
+	// if some child alias has returned partial results.
+	if len(indexErrors) > 0 || sr.Status.Failed > 0 {
+		sr = &SearchResult{
+			Status: sr.Status,
+		}
+		if sr.Status.Errors == nil {
+			sr.Status.Errors = make(map[string]error)
+		}
+		for indexName, indexErr := range indexErrors {
+			sr.Status.Errors[indexName] = indexErr
+			sr.Status.Total++
+			sr.Status.Failed++
+		}
+	}
+
+	return sr, nil
+}
+
+// hitsInCurrentPage returns the hits in the current page
+// using the From and Size parameters in the request
+func hitsInCurrentPage(req *SearchRequest, hits []*search.DocumentMatch) []*search.DocumentMatch {
+	sortFunc := req.SortFunc()
+	// sort all hits with the requested order
+	if len(req.Sort) > 0 {
+		sorter := newSearchHitSorter(req.Sort, hits)
+		sortFunc(sorter)
+	}
+	// now skip over the correct From
+	if req.From > 0 && len(hits) > req.From {
+		hits = hits[req.From:]
+	} else if req.From > 0 {
+		hits = search.DocumentMatchCollection{}
+	}
+	// now trim to the correct size
+	if req.Size > 0 && len(hits) > req.Size {
+		hits = hits[0:req.Size]
+	}
+	return hits
+}
+
 // MultiSearch executes a SearchRequest across multiple Index objects,
 // then merges the results.  The indexes must honor any ctx deadline.
-func MultiSearch(ctx context.Context, req *SearchRequest, indexes ...Index) (*SearchResult, error) {
+func MultiSearch(ctx context.Context, req *SearchRequest, preSearchData map[string]map[string]interface{}, indexes ...Index) (*SearchResult, error) {
 
 	searchStart := time.Now()
 	asyncResults := make(chan *asyncSearchResult, len(indexes))
@@ -466,7 +743,11 @@ func MultiSearch(ctx context.Context, req *SearchRequest, indexes ...Index) (*Se
 
 	waitGroup.Add(len(indexes))
 	for _, in := range indexes {
-		go searchChildIndex(in, createChildSearchRequest(req))
+		var payload map[string]interface{}
+		if preSearchData != nil {
+			payload = preSearchData[in.Name()]
+		}
+		go searchChildIndex(in, createChildSearchRequest(req, payload))
 	}
 
 	// on another go routine, close after finished
@@ -504,24 +785,7 @@ func MultiSearch(ctx context.Context, req *SearchRequest, indexes ...Index) (*Se
 		}
 	}
 
-	sortFunc := req.SortFunc()
-	// sort all hits with the requested order
-	if len(req.Sort) > 0 {
-		sorter := newSearchHitSorter(req.Sort, sr.Hits)
-		sortFunc(sorter)
-	}
-
-	// now skip over the correct From
-	if req.From > 0 && len(sr.Hits) > req.From {
-		sr.Hits = sr.Hits[req.From:]
-	} else if req.From > 0 {
-		sr.Hits = search.DocumentMatchCollection{}
-	}
-
-	// now trim to the correct size
-	if req.Size > 0 && len(sr.Hits) > req.Size {
-		sr.Hits = sr.Hits[0:req.Size]
-	}
+	sr.Hits = hitsInCurrentPage(req, sr.Hits)
 
 	// fix up facets
 	for name, fr := range req.Facets {
@@ -533,14 +797,16 @@ func MultiSearch(ctx context.Context, req *SearchRequest, indexes ...Index) (*Se
 		req.Sort.Reverse()
 		// resort using the original order
 		mhs := newSearchHitSorter(req.Sort, sr.Hits)
-		sortFunc(mhs)
+		req.SortFunc()(mhs)
 		// reset request
 		req.SearchBefore = req.SearchAfter
 		req.SearchAfter = nil
 	}
 
 	// fix up original request
-	sr.Request = req
+	if req.Explain {
+		sr.Request = req
+	}
 	searchDuration := time.Since(searchStart)
 	sr.Took = searchDuration
 
