@@ -304,7 +304,7 @@ func TestBytesWritten(t *testing.T) {
 	typeFieldMapping.DocValues = false
 	documentMapping.AddFieldMappingsAt("type", typeFieldMapping)
 
-	err = checkStatsOnIndexedBatch(tmpIndexPath, indexMapping, 37767)
+	err = checkStatsOnIndexedBatch(tmpIndexPath, indexMapping, 57273)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -313,7 +313,7 @@ func TestBytesWritten(t *testing.T) {
 	contentFieldMapping.Store = true
 	tmpIndexPath1 := createTmpIndexPath(t)
 
-	err := checkStatsOnIndexedBatch(tmpIndexPath1, indexMapping, 56582)
+	err := checkStatsOnIndexedBatch(tmpIndexPath1, indexMapping, 76069)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -323,7 +323,7 @@ func TestBytesWritten(t *testing.T) {
 	contentFieldMapping.IncludeInAll = true
 	tmpIndexPath2 := createTmpIndexPath(t)
 
-	err = checkStatsOnIndexedBatch(tmpIndexPath2, indexMapping, 44714)
+	err = checkStatsOnIndexedBatch(tmpIndexPath2, indexMapping, 68875)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -333,7 +333,7 @@ func TestBytesWritten(t *testing.T) {
 	contentFieldMapping.IncludeTermVectors = true
 	tmpIndexPath3 := createTmpIndexPath(t)
 
-	err = checkStatsOnIndexedBatch(tmpIndexPath3, indexMapping, 59479)
+	err = checkStatsOnIndexedBatch(tmpIndexPath3, indexMapping, 78985)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -343,7 +343,7 @@ func TestBytesWritten(t *testing.T) {
 	contentFieldMapping.DocValues = true
 	tmpIndexPath4 := createTmpIndexPath(t)
 
-	err = checkStatsOnIndexedBatch(tmpIndexPath4, indexMapping, 44722)
+	err = checkStatsOnIndexedBatch(tmpIndexPath4, indexMapping, 64228)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -401,8 +401,8 @@ func TestBytesRead(t *testing.T) {
 	}
 	stats, _ := idx.StatsMap()["index"].(map[string]interface{})
 	prevBytesRead, _ := stats["num_bytes_read_at_query_time"].(uint64)
-	if prevBytesRead != 36066 && res.Cost == prevBytesRead {
-		t.Fatalf("expected bytes read for query string 32349, got %v",
+	if prevBytesRead != 21639 && res.Cost == prevBytesRead {
+		t.Fatalf("expected bytes read for query string 21639, got %v",
 			prevBytesRead)
 	}
 
@@ -504,33 +504,6 @@ func TestBytesRead(t *testing.T) {
 	}
 }
 
-func getBatchFromData(idx Index, fileName string) (*Batch, error) {
-	pwd, err := os.Getwd()
-	if err != nil {
-		return nil, err
-	}
-	path := filepath.Join(pwd, "data", "test", fileName)
-	batch := idx.NewBatch()
-	var dataset []map[string]interface{}
-	fileContent, err := os.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-	err = json.Unmarshal(fileContent, &dataset)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, doc := range dataset {
-		err = batch.Index(fmt.Sprintf("%d", doc["id"]), doc)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return batch, err
-}
-
 func TestBytesReadStored(t *testing.T) {
 	tmpIndexPath := createTmpIndexPath(t)
 	defer cleanupTmpIndexPath(t, tmpIndexPath)
@@ -580,8 +553,8 @@ func TestBytesReadStored(t *testing.T) {
 
 	stats, _ := idx.StatsMap()["index"].(map[string]interface{})
 	bytesRead, _ := stats["num_bytes_read_at_query_time"].(uint64)
-	if bytesRead != 25928 && bytesRead == res.Cost {
-		t.Fatalf("expected the bytes read stat to be around 25928, got %v", bytesRead)
+	if bytesRead != 11501 && bytesRead == res.Cost {
+		t.Fatalf("expected the bytes read stat to be around 11501, got %v", bytesRead)
 	}
 	prevBytesRead := bytesRead
 
@@ -651,8 +624,8 @@ func TestBytesReadStored(t *testing.T) {
 
 	stats, _ = idx1.StatsMap()["index"].(map[string]interface{})
 	bytesRead, _ = stats["num_bytes_read_at_query_time"].(uint64)
-	if bytesRead != 18114 && bytesRead == res.Cost {
-		t.Fatalf("expected the bytes read stat to be around 18114, got %v", bytesRead)
+	if bytesRead != 3687 && bytesRead == res.Cost {
+		t.Fatalf("expected the bytes read stat to be around 3687, got %v", bytesRead)
 	}
 	prevBytesRead = bytesRead
 
@@ -678,6 +651,33 @@ func TestBytesReadStored(t *testing.T) {
 	if bytesRead-prevBytesRead != 42 && bytesRead-prevBytesRead == res.Cost {
 		t.Fatalf("expected the bytes read stat to be around 42, got %v", bytesRead-prevBytesRead)
 	}
+}
+
+func getBatchFromData(idx Index, fileName string) (*Batch, error) {
+	pwd, err := os.Getwd()
+	if err != nil {
+		return nil, err
+	}
+	path := filepath.Join(pwd, "data", "test", fileName)
+	batch := idx.NewBatch()
+	var dataset []map[string]interface{}
+	fileContent, err := os.ReadFile(path)
+	if err != nil {
+		return nil, err
+	}
+	err = json.Unmarshal(fileContent, &dataset)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, doc := range dataset {
+		err = batch.Index(fmt.Sprintf("%d", doc["id"]), doc)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return batch, err
 }
 
 func TestIndexCreateNewOverExisting(t *testing.T) {
