@@ -35,7 +35,7 @@ func init() {
 
 type ConjunctionSearcher struct {
 	indexReader index.IndexReader
-	searchers   OrderedSearcherList
+	searchers   []search.Searcher
 	queryNorm   float64
 	currs       []*search.DocumentMatch
 	maxIDIdx    int
@@ -88,6 +88,20 @@ func NewConjunctionSearcher(ctx context.Context, indexReader index.IndexReader,
 	return &rv, nil
 }
 
+func (s *ConjunctionSearcher) computeQueryNorm() {
+	// first calculate sum of squared weights
+	sumOfSquaredWeights := 0.0
+	for _, searcher := range s.searchers {
+		sumOfSquaredWeights += searcher.Weight()
+	}
+	// now compute query norm from this
+	s.queryNorm = 1.0 / math.Sqrt(sumOfSquaredWeights)
+	// finally tell all the downstream searchers the norm
+	for _, searcher := range s.searchers {
+		searcher.SetQueryNorm(s.queryNorm)
+	}
+}
+
 func (s *ConjunctionSearcher) Size() int {
 	sizeInBytes := reflectStaticSizeConjunctionSearcher + size.SizeOfPtr +
 		s.scorer.Size()
@@ -103,20 +117,6 @@ func (s *ConjunctionSearcher) Size() int {
 	}
 
 	return sizeInBytes
-}
-
-func (s *ConjunctionSearcher) computeQueryNorm() {
-	// first calculate sum of squared weights
-	sumOfSquaredWeights := 0.0
-	for _, searcher := range s.searchers {
-		sumOfSquaredWeights += searcher.Weight()
-	}
-	// now compute query norm from this
-	s.queryNorm = 1.0 / math.Sqrt(sumOfSquaredWeights)
-	// finally tell all the downstream searchers the norm
-	for _, searcher := range s.searchers {
-		searcher.SetQueryNorm(s.queryNorm)
-	}
 }
 
 func (s *ConjunctionSearcher) initSearchers(ctx *search.SearchContext) error {
