@@ -39,22 +39,25 @@ func init() {
 }
 
 type SearcherCurr struct {
-	searcher search.Searcher
-	curr     *search.DocumentMatch
+	searcher    search.Searcher
+	curr        *search.DocumentMatch
+	matchingIdx int
 }
 
 type DisjunctionHeapSearcher struct {
 	indexReader index.IndexReader
 
-	numSearchers int
-	scorer       *scorer.DisjunctionQueryScorer
-	min          int
-	queryNorm    float64
-	initialized  bool
-	searchers    []search.Searcher
-	heap         []*SearcherCurr
+	numSearchers           int
+	scorer                 *scorer.DisjunctionQueryScorer
+	min                    int
+	queryNorm              float64
+	retrieveScoreBreakdown bool
+	initialized            bool
+	searchers              []search.Searcher
+	heap                   []*SearcherCurr
 
 	matching      []*search.DocumentMatch
+	matchingIdxs  []int
 	matchingCurrs []*SearcherCurr
 
 	bytesRead uint64
@@ -67,20 +70,40 @@ func newDisjunctionHeapSearcher(ctx context.Context, indexReader index.IndexRead
 	if limit && tooManyClauses(len(searchers)) {
 		return nil, tooManyClausesErr("", len(searchers))
 	}
+	var retrieveScoreBreakdown bool
+	if ctx != nil {
+		retrieveScoreBreakdown, _ = ctx.Value(search.IncludeScoreBreakdownKey).(bool)
+	}
 
 	// build our searcher
 	rv := DisjunctionHeapSearcher{
-		indexReader:   indexReader,
-		searchers:     searchers,
-		numSearchers:  len(searchers),
-		scorer:        scorer.NewDisjunctionQueryScorer(options),
-		min:           int(min),
-		matching:      make([]*search.DocumentMatch, len(searchers)),
-		matchingCurrs: make([]*SearcherCurr, len(searchers)),
-		heap:          make([]*SearcherCurr, 0, len(searchers)),
+		indexReader:            indexReader,
+		searchers:              searchers,
+		numSearchers:           len(searchers),
+		scorer:                 scorer.NewDisjunctionQueryScorer(options),
+		min:                    int(min),
+		matching:               make([]*search.DocumentMatch, len(searchers)),
+		matchingCurrs:          make([]*SearcherCurr, len(searchers)),
+		matchingIdxs:           make([]int, len(searchers)),
+		retrieveScoreBreakdown: retrieveScoreBreakdown,
+		heap:                   make([]*SearcherCurr, 0, len(searchers)),
 	}
 	rv.computeQueryNorm()
 	return &rv, nil
+}
+
+func (s *DisjunctionHeapSearcher) computeQueryNorm() {
+	// first calculate sum of squared weights
+	sumOfSquaredWeights := 0.0
+	for _, searcher := range s.searchers {
+		sumOfSquaredWeights += searcher.Weight()
+	}
+	// now compute query norm from this
+	s.queryNorm = 1.0 / math.Sqrt(sumOfSquaredWeights)
+	// finally tell all the downstream searchers the norm
+	for _, searcher := range s.searchers {
+		searcher.SetQueryNorm(s.queryNorm)
+	}
 }
 
 func (s *DisjunctionHeapSearcher) Size() int {
@@ -101,22 +124,9 @@ func (s *DisjunctionHeapSearcher) Size() int {
 	// since searchers and document matches already counted above
 	sizeInBytes += len(s.matchingCurrs) * reflectStaticSizeSearcherCurr
 	sizeInBytes += len(s.heap) * reflectStaticSizeSearcherCurr
+	sizeInBytes += len(s.matchingIdxs) * size.SizeOfInt
 
 	return sizeInBytes
-}
-
-func (s *DisjunctionHeapSearcher) computeQueryNorm() {
-	// first calculate sum of squared weights
-	sumOfSquaredWeights := 0.0
-	for _, searcher := range s.searchers {
-		sumOfSquaredWeights += searcher.Weight()
-	}
-	// now compute query norm from this
-	s.queryNorm = 1.0 / math.Sqrt(sumOfSquaredWeights)
-	// finally tell all the downstream searchers the norm
-	for _, searcher := range s.searchers {
-		searcher.SetQueryNorm(s.queryNorm)
-	}
 }
 
 func (s *DisjunctionHeapSearcher) initSearchers(ctx *search.SearchContext) error {
@@ -132,6 +142,7 @@ func (s *DisjunctionHeapSearcher) initSearchers(ctx *search.SearchContext) error
 		if curr != nil {
 			block[i].searcher = searcher
 			block[i].curr = curr
+			block[i].matchingIdx = i
 			heap.Push(s, &block[i])
 		}
 	}
@@ -147,6 +158,7 @@ func (s *DisjunctionHeapSearcher) initSearchers(ctx *search.SearchContext) error
 func (s *DisjunctionHeapSearcher) updateMatches() error {
 	matching := s.matching[:0]
 	matchingCurrs := s.matchingCurrs[:0]
+	matchingIdxs := s.matchingIdxs[:0]
 
 	if len(s.heap) > 0 {
 
@@ -154,17 +166,20 @@ func (s *DisjunctionHeapSearcher) updateMatches() error {
 		next := heap.Pop(s).(*SearcherCurr)
 		matching = append(matching, next.curr)
 		matchingCurrs = append(matchingCurrs, next)
+		matchingIdxs = append(matchingIdxs, next.matchingIdx)
 
 		// now as long as top of heap matches, keep popping
 		for len(s.heap) > 0 && bytes.Compare(next.curr.IndexInternalID, s.heap[0].curr.IndexInternalID) == 0 {
 			next = heap.Pop(s).(*SearcherCurr)
 			matching = append(matching, next.curr)
 			matchingCurrs = append(matchingCurrs, next)
+			matchingIdxs = append(matchingIdxs, next.matchingIdx)
 		}
 	}
 
 	s.matching = matching
 	s.matchingCurrs = matchingCurrs
+	s.matchingIdxs = matchingIdxs
 
 	return nil
 }
@@ -197,10 +212,16 @@ func (s *DisjunctionHeapSearcher) Next(ctx *search.SearchContext) (
 	for !found && len(s.matching) > 0 {
 		if len(s.matching) >= s.min {
 			found = true
-			partialMatch := len(s.matching) != len(s.searchers)
-			// score this match
-			rv = s.scorer.Score(ctx, s.matching, len(s.matching), s.numSearchers)
-			rv.PartialMatch = partialMatch
+			if s.retrieveScoreBreakdown {
+				// just return score and expl breakdown here, since it is a disjunction over knn searchers,
+				// and the final score and expl is calculated in the knn collector
+				rv = s.scorer.ScoreAndExplBreakdown(ctx, s.matching, s.matchingIdxs, nil, s.numSearchers)
+			} else {
+				// score this match
+				partialMatch := len(s.matching) != len(s.searchers)
+				rv = s.scorer.Score(ctx, s.matching, len(s.matching), s.numSearchers)
+				rv.PartialMatch = partialMatch
+			}
 		}
 
 		// invoke next on all the matching searchers
